Address cuda.core coverage test review comments
